+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Added light theme for the application
 - Added App icon, set app name to Biclou
 - Added Location features
+- Added MainActivity navigation